# docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "PR News Backend API",
		"version": "1.0.0",
		"description": "Pull Request解説記事生成サービスのバックエンドAPI仕様書。API設計ガイドラインに基づき、Hono + Zod + Hono Client を用いたREST APIの設計方針を反映しています。\n共通レスポンスとして ApiResponse<T> を採用し、エラーコードは ErrorCode として列挙型で管理します。"
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Firebase IDトークンを Bearer トークンとして指定します。"
			}
		},
		"schemas": {
			"ErrorCode": {
				"type": "string",
				"description": "アプリケーション固有のエラーコードの列挙型",
				"enum": [
					"VALIDATION_ERROR",
					"UNAUTHENTICATED",
					"FORBIDDEN",
					"NOT_FOUND",
					"INTERNAL_SERVER_ERROR"
				],
				"example": "NOT_FOUND"
			},
			"ErrorObject": {
				"type": "object",
				"description": "エラーレスポンスの詳細情報",
				"properties": {
					"code": {
						"$ref": "#/components/schemas/ErrorCode"
					},
					"details": {
						"type": "object",
						"nullable": true,
						"description": "エラーの詳細情報（バリデーションエラーの場合、フィールドごとのエラー内容など）",
						"additionalProperties": true,
						"example": [
							{
								"field": "email",
								"reason": "正しいメールアドレス形式で入力してください"
							}
						]
					}
				},
				"required": ["code"]
			},
			"User": {
				"type": "object",
				"description": "ユーザー情報を表すスキーマ",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"description": "ユーザーID (UUID形式)"
					},
					"githubUserId": {
						"type": "integer",
						"description": "GitHub ユーザーID (整数、正の数)",
						"minimum": 1
					},
					"githubUsername": {
						"type": "string",
						"description": "GitHub ユーザー名 (必須)",
						"minLength": 1
					},
					"githubDisplayName": {
						"type": "string",
						"nullable": true,
						"description": "GitHub 表示名"
					},
					"email": {
						"type": "string",
						"format": "email",
						"nullable": true,
						"description": "メールアドレス (正しいメールアドレス形式)"
					},
					"avatarUrl": {
						"type": "string",
						"format": "url",
						"nullable": true,
						"description": "アバター画像のURL (正しいURL形式)"
					},
					"language": {
						"type": "string",
						"description": "ユーザーの希望言語コード (例: ja, en)",
						"minLength": 2,
						"maxLength": 2,
						"default": "ja"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "作成日時 (正しい日時形式)",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "更新日時 (正しい日時形式)",
						"nullable": true
					}
				},
				"required": ["id", "githubUserId", "githubUsername", "language"]
			},
...
			"LogoutResponseData": {
				"type": "object",
				"description": "ログアウト成功時の空のデータオブジェクト"
			},
			"FavoriteRepository": {
				"type": "object",
				"description": "お気に入りリポジトリ情報",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"description": "お気に入り登録ID"
					},
					"userId": {
						"type": "string",
						"format": "uuid",
						"description": "ユーザーID"
					},
					"githubRepoId": {
						"type": "integer",
						"description": "GitHubリポジトリの数値ID"
					},
					"repositoryFullName": {
						"type": "string",
						"description": "リポジトリフルネーム (owner/repo)"
					},
					"owner": { "type": "string", "description": "オーナー名" },
					"repo": { "type": "string", "description": "リポジトリ名" },
					"registeredAt": {
						"type": "string",
						"format": "date-time",
						"description": "登録日時"
					}
				},
				"required": [
					"id",
					"userId",
					"githubRepoId",
					"repositoryFullName",
					"owner",
					"repo",
					"registeredAt"
				]
			},
...
			"LikedArticleInfo": {
				"type": "object",
				"description": "いいねした記事の要約情報",
				"properties": {
					"articleId": { "type": "string", "format": "uuid" },
					"languageCode": { "type": "string", "pattern": "^[a-z]{2}$" },
					"likedAt": { "type": "string", "format": "date-time" },
					"aiGeneratedTitle": { "type": "string" },
					"repositoryFullName": { "type": "string" },
					"prNumber": { "type": "integer" }
				},
				"required": [
					"articleId",
					"languageCode",
					"likedAt",
					"aiGeneratedTitle",
					"repositoryFullName",
					"prNumber"
				]
			},
			"RankedArticleInfo": {
				"type": "object",
				"description": "ランキングされた記事の情報",
				"properties": {
					"rank": { "type": "integer" },
					"articleId": { "type": "string", "format": "uuid" },
					"languageCode": { "type": "string", "pattern": "^[a-z]{2}$" },
					"aiGeneratedTitle": { "type": "string" },
					"repositoryFullName": { "type": "string" },
					"prNumber": { "type": "integer" },
					"likeCount": { "type": "integer" }
				},
				"required": [
					"rank",
					"articleId",
					"languageCode",
					"aiGeneratedTitle",
					"repositoryFullName",
					"prNumber",
					"likeCount"
				]
			},
			"Pagination": {
				"type": "object",
				"description": "ページネーション情報",
				"properties": {
					"totalItems": { "type": "integer" },
					"limit": { "type": "integer" },
					"offset": { "type": "integer" }
				},
				"required": ["totalItems", "limit", "offset"]
			}
		}
	},
	"paths": {
...
		"/users/me": {
			"get": {
				"summary": "現在のユーザー情報を取得",
				"description": "認証されている現在のユーザーのプロファイル情報を取得します。フロントエンドのユーザーダッシュボード初期読み込みなどに使用されます。",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"tags": ["User & Auth"],
				"responses": {
					"200": {
						"description": "ユーザー情報の取得成功",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/User"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "data"]
								},
								"example": {
									"success": true,
									"data": {
										"id": "5f8c9c1d-df89-4294-8218-a303ef525b6c",
										"githubUsername": "octocat",
										"avatarUrl": "https://avatars.githubusercontent.com/u/583231?v=4"
									}
								}
							}
						}
					},
					"401": {
						"description": "認証エラー (トークン無効など)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": false
										},
										"error": {
											"$ref": "#/components/schemas/ErrorObject"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "error"]
								},
								"example": {
									"success": false,
									"error": {
										"code": "UNAUTHENTICATED"
									},
									"message": "認証トークンが無効です。再度ログインしてください。"
								}
							}
						}
					},
					"500": {
						"description": "サーバー内部エラー",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": false
										},
										"error": {
											"$ref": "#/components/schemas/ErrorObject"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "error"]
								},
								"example": {
									"success": false,
									"error": {
										"code": "INTERNAL_SERVER_ERROR"
									},
									"message": "サーバー内部でエラーが発生しました。"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "お気に入りリポジトリを登録",
				"description": "指定されたGitHubリポジトリをユーザーのお気に入りとして登録します。",
				"tags": ["Favorites"],
				"security": [{ "bearerAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"owner": {
										"type": "string",
										"description": "リポジトリのオーナー名"
									},
									"repo": { "type": "string", "description": "リポジトリ名" }
								},
								"required": ["owner", "repo"]
							},
							"example": { "owner": "octocat", "repo": "Hello-World" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "リポジトリは既にお気に入り登録済み (既存の情報を返却)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean", "example": true },
										"data": {
											"$ref": "#/components/schemas/FavoriteRepository"
										},
										"message": { "type": "string", "nullable": true }
									}
								}
							}
						}
					},
					"201": {
						"description": "お気に入りリポジトリの登録成功 (新規登録)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean", "example": true },
										"data": {
											"$ref": "#/components/schemas/FavoriteRepository"
										},
										"message": { "type": "string", "nullable": true }
									}
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/BadRequest" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"404": {
						"description": "指定されたGitHubリポジトリが見つからない",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean", "example": false },
										"error": { "$ref": "#/components/schemas/ErrorObject" },
										"message": { "type": "string", "nullable": true }
									},
									"example": {
										"success": false,
										"error": { "code": "GITHUB_REPO_NOT_FOUND" },
										"message": "指定されたGitHubリポジトリが見つかりません。"
									}
								}
							}
						}
					},
					"500": { "$ref": "#/components/responses/InternalServerError" }
				}
			}
		},
		"/auth/logout": {
			"post": {
				"summary": "ログアウト処理",
				"description": "現在のユーザーセッションを無効化し、ログアウトします。サーバー側でAuthSessionのrevokedAtを更新します。",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"tags": ["User & Auth"],
				"responses": {
					"200": {
						"description": "ログアウト成功",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/LogoutResponseData"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "data"]
								},
								"example": {
									"success": true,
									"data": {},
									"message": "Logged out successfully"
								}
							}
						}
					},
					"401": {
						"description": "認証エラー",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": false
										},
										"error": {
											"$ref": "#/components/schemas/ErrorObject"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "error"]
								},
								"example": {
									"success": false,
									"error": {
										"code": "UNAUTHENTICATED"
									},
									"message": "認証されていません。"
								}
							}
						}
					},
					"500": {
						"description": "サーバー内部エラー",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": false
										},
										"error": {
											"$ref": "#/components/schemas/ErrorObject"
										},
										"message": {
											"type": "string",
											"nullable": true
										}
									},
									"required": ["success", "error"]
								},
								"example": {
									"success": false,
									"error": {
										"code": "INTERNAL_SERVER_ERROR"
									},
									"message": "サーバー内部でエラーが発生しました。"
								}
							}
						}
					}
				}
			}
		},
		"/auth/signup": {
			"post": {
				"summary": "ユーザーサインアップ",
				"description": "新規ユーザー登録。language（2文字コード、省略時ja）を指定可能。",
				"security": [{ "bearerAuth": [] }],
				"tags": ["User & Auth"],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"language": {
										"type": "string",
										"description": "希望言語コード (例: ja, en)",
										"minLength": 2,
										"maxLength": 2,
										"default": "ja"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "ユーザー作成成功",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"409": {
						"description": "既存ユーザー",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorObject"
								}
							}
						}
					}
				}
			}
		},
...
		"/users/me/liked-articles": {
			"get": {
				"summary": "自分が「いいね」した記事の一覧を取得",
				"description": "認証ユーザーが過去に「いいね」したAI解説記事の一覧を、絞り込み、ページネーション、ソート指定付きで取得します。",
				"tags": ["User & Auth", "Likes"],
				"security": [{ "bearerAuth": [] }],
				"parameters": [
					{
						"name": "lang",
						"in": "query",
						"required": false,
						"description": "特定の言語でいいねした記事に絞り込む (例: ja, en)",
						"schema": { "type": "string", "pattern": "^[a-z]{2}$" }
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "取得件数 (デフォルトはサーバー側で設定)",
						"schema": { "type": "integer", "minimum": 1, "default": 10 }
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"description": "開始位置 (デフォルトは0)",
						"schema": { "type": "integer", "minimum": 0, "default": 0 }
					},
					{
						"name": "sort",
						"in": "query",
						"required": false,
						"description": "ソート順 (例: likedAt_desc, likedAt_asc)",
						"schema": {
							"type": "string",
							"enum": ["likedAt_desc", "likedAt_asc"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "自分がいいねした記事一覧の取得成功",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean", "example": true },
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/LikedArticleInfo"
											}
										},
										"message": { "type": "string", "nullable": true },
										"pagination": {
											"type": "object",
											"properties": {
												"totalItems": { "type": "integer" },
												"limit": { "type": "integer" },
												"offset": { "type": "integer" }
											}
										}
									}
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/BadRequest" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"500": { "$ref": "#/components/responses/InternalServerError" }
				}
			}
		},
		"/ranking/articles/likes": {
			"get": {
				"summary": "いいね数に基づいた記事ランキングを取得",
				"description": "AI解説記事のいいね数ランキングを、期間や言語で絞り込んで取得します。",
				"tags": ["Ranking"],
				"parameters": [
					{
						"name": "period",
						"in": "query",
						"required": false,
						"description": "集計期間",
						"schema": {
							"type": "string",
							"enum": ["weekly", "monthly", "all"],
							"default": "all"
						}
					},
					{
						"name": "language",
						"in": "query",
						"required": false,
						"description": "言語での絞り込み (ja, en, all)",
						"schema": {
							"type": "string",
							"pattern": "^[a-z]{2}|all$",
							"default": "all"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "取得件数",
						"schema": { "type": "integer", "minimum": 1, "default": 10 }
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"description": "開始位置",
						"schema": { "type": "integer", "minimum": 0, "default": 0 }
					}
				],
				"responses": {
					"200": {
						"description": "記事ランキングの取得成功",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean", "example": true },
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/RankedArticleInfo"
											}
										},
										"pagination": { "$ref": "#/components/schemas/Pagination" },
										"message": { "type": "string", "nullable": true }
									}
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/BadRequest" },
					"500": { "$ref": "#/components/responses/InternalServerError" }
				}
			}
		}
	},
...
}