# packages/common/src/http/errorCodes.ts
export const ErrorCode = {
	VALIDATION_ERROR: "VALIDATION_ERROR",
	UNAUTHENTICATED: "UNAUTHENTICATED",
	FORBIDDEN: "FORBIDDEN",
	NOT_FOUND: "NOT_FOUND",
	INTERNAL_SERVER_ERROR: "INTERNAL_SERVER_ERROR",
	GITHUB_REPO_NOT_FOUND: "GITHUB_REPO_NOT_FOUND",
	ARTICLE_NOT_FOUND: "ARTICLE_NOT_FOUND",
} as const;

export type ErrorCode = (typeof ErrorCode)[keyof typeof ErrorCode];